+++ app/changelog.md
@@ -42,3 +42,4 @@ _ Update to Angular Material 0.10.0
 - Restyled menu scrollbars for Windows IE and Chrome
 - Added FAB Speed Dial & FAB Toolbar element example pages
 - Added Fab Speed Dial to General Blog Widget
+- Added Menu example elements page